help: don't show factor as part of unit
bugfix for commit e494afabde429ec27eabe9d457943c02f8126637

the factor is applied internally by cantools
and is therefore irrelevant for the end user.
But it was displayed as part of the unit
so the end user could not know whether
they needed to scale the values themselves.
From now on it is no longer displayed by default.
You can still display it using --bits.